# Revised SND DROP and SND pacing logic