### Sending Priority

In the live streaming configuration with TL Packet Drop enabled original packet must have priority over packet retransmission because:
- The earlier the latest original packet reaches the receiver, the better the end-to-end latency constraint is kept as the receiver can detect too late packets earlier.
- If original packets are accumulating in the sender buffer, there is a network congestion, and likely most of the packet retransmissions won't reach the receiver. Therefore it is better to prioritize a more up-to-date payload, preserve the end-to-end latency and keep the jitter low. Once congestion is over, the streaming can be restored quicker.
- In-order but belated packets have a negative impact on jitter and end-to-end latency.

Proposed sending priority (for Live Configuration Only):
1. Retransmit a lost packet only if there is no original packet available for sending (which happens when the current sending rate < MaxBW).
2. Otherwise check if FEC packet can be sent.
3. Otherwise send the next original packet (first time sending).

## Extracted Changes

### Prefer Using Source Time if Provided Everywhere (Moved to #2185)

If a source time is provided, it must be considered as the origin time of a message. Meaning that buffer timespan calculation and the TL packet drop logic must rely on this time instead.
Thus there is no need to save two timestamps (`m_tsOriginTime` and `m_llSourceTime_us`). Just use `m_tsOriginTime` everywhere.

### Refactored `CUDT::packData()` (Moved to #2229)

The logic which extracts an original (not ever sent yet_ packet from the sender buffer has been moved to a dedicated function `CUDT::packUniqueData(..)`. The encryption order has been changed.

- [x] Verify sending an encrypted does not have issues.

### Fix SND Drop Logic (Moved to #2230)

Sender dropping logic is applied if the sender buffer timespan exceeds `drop_threshold_ms`.
Instead, the delay of the oldest packet in the buffer should be used (`Tnow - OldestPacketTS`).
This also correlates with the usage of the source time if provided to calculate the timestamp, which was not the case previously.

As before, only packets older than `Tnow - drop_threshold_ms ` are dropped by the sender.

See issue #898, #1910.

### Don't Suppress SND Pacing (Moved to #2232)

Don't allow to force sending if `checkNeedDrop` signals congestion.
Addresses #713.