@maxsharabayko 2022-01-28
## Test Environment

- Two CentOS 7 test machines.
- **Congestion**: rate limit (LanForge).
- **Sending rate (source bitrate)**: 7 Mbps.
- **Network throughput:**
   - 8 Mbps (+14%);
   - 5 Mbps (-28%) from 20th to 30th second. 
- **SRT Max BW limit**: 8 Mbps (+14% from source bitrate).
- **Packet loss**: No (only by rate limit).
- **RTT**: 40 ms.
- **SRT Latency**: 150 ms.

```shell
(receiver)
./srt-xtransmit receive srt://:4200 --statsfile stats-rcv.csv --statsfreq 1s --metricsfile stats-metrics.csv \
                --metricsfreq 1s -v --enable-metrics

(sender)
maxbw=1000000
sendrate=7Mbps
./srt-xtransmit generate "srt://ip:4200?maxbw=$maxbw&latency=150" --sendrate $rate --duration 60s \
                --statsfile stats-snd.csv --statsfreq 100ms --enable-metrics -v

```

| Vertsion | pktRecvTotal | pktRcvDropTotal | pktRecvBelatedTotal | pktRecvAppTotal | pktMissingAppTotal |
| -------- | ------------| ------- | ------- | ----- | ----- |
| this PR | 37842 | 2229 | 196 | **36900** | **2229** |
| master | 38414 | 2397 | 935 | **36767** | **2397** |

_Note_. `pktRecvAppTotal` and `pktMissingAppTotal` are calculated by the receiving application (srt-xtransmit, `stats-metrics.csv`). Other statistics are calculated by SRT.

### SRT master after PR #2232

### This PR

⚠️  100 ms stats interval on the sender.

<img src="https://user-images.githubusercontent.com/12700120/151484437-9cb5fa5d-a9f1-4a55-b7b1-b8425d9f374f.png" width="700">

<img src="https://user-images.githubusercontent.com/12700120/151484394-a5954a83-cc13-4ffe-82bb-0f1750c6be80.png" width="700">

<img src="https://user-images.githubusercontent.com/12700120/151484454-7b7ebe60-81ee-4487-8350-0a6c2b50b4c4.png" width="700">

<img src="https://user-images.githubusercontent.com/12700120/151485714-9c0d85bc-0cd3-44df-993a-700252050f28.png" width="700">

<img src="https://user-images.githubusercontent.com/12700120/151485753-e063a5b3-0645-4cd9-805d-fb135d3eabc0.png" width="700">


### SRT master after PR #2232

⚠️  100 ms stats interval on the sender.

<img src="https://user-images.githubusercontent.com/12700120/150957360-504f549f-a88d-4b92-a1d6-8cb75b9c0bbf.png" width="700">

<img src="https://user-images.githubusercontent.com/12700120/150957394-58e9675d-1e7e-46e0-813e-9556ef96815f.png" width="700">

<img src="https://user-images.githubusercontent.com/12700120/150957485-c4eca1fd-cf41-4fb4-b1af-d92356392761.png" width="700">

<img src="https://user-images.githubusercontent.com/12700120/151485894-0cd4c619-0866-4692-8afb-cb3dcfa0a4cd.png" width="700">

<img src="https://user-images.githubusercontent.com/12700120/151485913-46d16c41-b626-421d-a69a-f800f5a23678.png" width="700">